requirements.txt
numpy==1.21.2
openai==0.27.0
python-dotenv==0.19.0
requests==2.26.0
scikit-learn==1.0
aiofiles==23.1.0
fastapi
uvicorn
jinja2